Move contact details, nominated contacts into CHEDPP package
