+++ trade-mapping-to-ipaffs/src/main/java/uk/gov/defra/stw/mapping/toipaffs/chedpp/ChedppPartOneMapper.java
@@ -13,11 +13,9 @@ import uk.gov.defra.stw.mapping.dto.SpsExchangedDocument;
 import uk.gov.defra.stw.mapping.toipaffs.Mapper;
 import uk.gov.defra.stw.mapping.toipaffs.common.ArrivalDateMapper;
 import uk.gov.defra.stw.mapping.toipaffs.common.ArrivalTimeMapper;
-import uk.gov.defra.stw.mapping.toipaffs.common.ContactDetailsMapper;
 import uk.gov.defra.stw.mapping.toipaffs.common.DepartureDateMapper;
 import uk.gov.defra.stw.mapping.toipaffs.common.DepartureTimeMapper;
 import uk.gov.defra.stw.mapping.toipaffs.common.EconomicOperatorMapper;
-import uk.gov.defra.stw.mapping.toipaffs.common.NominatedContactsMapper;
 import uk.gov.defra.stw.mapping.toipaffs.common.PointOfEntryMapper;
 import uk.gov.defra.stw.mapping.toipaffs.common.SealsContainersMapper;
 import uk.gov.defra.stw.mapping.toipaffs.common.TransportToBcpQuestionMapper;
