+++ trade-mapping-to-ipaffs/src/main/java/uk/gov/defra/stw/mapping/toipaffs/chedpp/ContactDetailsMapper.java
@@ -1,4 +1,4 @@
-package uk.gov.defra.stw.mapping.toipaffs.common;
+package uk.gov.defra.stw.mapping.toipaffs.chedpp;
 
 import java.util.List;
 import org.springframework.stereotype.Component;
